+++ CHANGELOG.markdown
@@ -1,5 +1,8 @@
 # Changelog
 
+## social-wallet-api 0.10.0
+* Added pagination to transaction/list
+
 ## social-wallet-api 0.9.4
 * Jetty host can be picked up as a env var